+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 [Semantic Versioning].
 
 ## [Unreleased]
+
+## [0.1.4] - 2017-09-26
 ### Added
 - `PartialOrd` blanket `impl` for `MemOrdered` works for any type that the inner
   type implements `MemOrd` for.
@@ -40,7 +42,8 @@ Initial release
 [Keep a Changelog]:    http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/nvzqz/mem-cmp-rs/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/nvzqz/mem-cmp-rs/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/nvzqz/mem-cmp-rs/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/nvzqz/mem-cmp-rs/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/nvzqz/mem-cmp-rs/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/nvzqz/mem-cmp-rs/compare/v0.1.0...v0.1.1